Adding the demo app header back into the document flow
When testing PR #7669 to make sure that the sandpack demos of cord.com are not broken, I noticed that all demos in sandpack had a huge margin, due to the new header (which is removed in cord.com). This margin existed previously in the demos, but it was not noticeable as the header was a lot smaller.

We already have to special-case margins to run the demo apps in the docs. Ideally we would want to avoid adding a another special case for sandpack. The main issue was that the header had `position: absolute` which took it out of the document flow and forced us to add a margin to the rest of the demo app.

This PR fixes this issue by making the header take the default `position: static` CSS, and correcting all margins from the demo apps - making sure all relevant framing styles are no longer in the #root element. In that way, we don't need to special case anything else, other than the documentation website.

Test Plan:
Tested all demo apps: locally, in the local docs and in the local v5 cord.com.


Reviewers: Netceer, elenaperezrioja, maylynn-ng, km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7820

monorepo-commit: 50e78dddd34d3723cbe90c2682684e842bb41bab