Demo app dropdown redesign (3/3): Adding the new component dropdown
We have redesigned the "How it's built" dropdown in the demo apps in order to make it more discoverable. Instead of finding the dropdown button, then clicking it and eventually showing the list of components used, now all components used are immediately visible when opening a demo application.

Subsequent PRs have prepared each demo app to be framed correctly for the redesign. This PR implements the final step of the plan outlined in #7691, which refactors the demo app header for the local applications as well as the documentation pages. It also corrects margins as necessary in each app (both locally and in the docs), as the header now has a larger height. Finally, it adds all icons needed for the component buttons.

Test Plan:
Tested both locally and in the local documentation website. Opened each demo application and clicked into the component and API links. All directed to the correct pages and highlighted components correctly, where relevant.

Also tested the sandpack build in cord.com to make sure that nothing breaks with the new margins or any of the changes in the headers.


Reviewers: flooey, maylynn-ng, elenaperezrioja, km-nur

Reviewed By: km-nur, elenaperezrioja

Pull Request: https://github.com/getcord/monorepo/pull/7669

monorepo-commit: e83a8f1f95bbd56900bbe80c950eecac3c2d225a