QD: Correcting composer padding in the document demo
The padding in the composer menu in the document demo was not quite right, as the buttons would touch the border of the composer. This happens because the padding was reset to 0 in all sides.

This PR corrects this by only resetting the top and bottom padding to 0, while keeping the existing left and right.

Test Plan:
Before:
<img width="204" alt="Screenshot 2024-01-18 at 15 36 46" src="https://github.com/getcord/monorepo/assets/102963329/78e9845d-dfc6-4946-8a86-f83c42ad4ead">


After:
<img width="214" alt="Screenshot 2024-01-18 at 15 36 33" src="https://github.com/getcord/monorepo/assets/102963329/46df25ff-d004-4eab-acbd-60839d5eee0b">



Reviewers: km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7922

monorepo-commit: 47d998a4fbb8b87e04ccc79a41bb31a253f3fdc3