Document demo: Fix clicking on the Thread composer bug
When clicking on an open thread in the document demo, the thread is removed from focus. So it would close again. This is a bug as we cannot really interact with a thread to add a reaction of re-focus the composer.

This bug was introduced in #7696 when framing the document demo. Setting a z-index to the open thread, did not take into account the click underlay, used for understanding when a user clicked outside the open-thread. This caused the open thread to be positioned under the click underlay in the stacking context, which in turn caused this behavior.

This is fixed by removing the set z-index and letting the threads get a `z-index: auto`, which correctly positions them in the stacking context.

Note: I also updated the sandpack file, to make sure these fixes are reflected in cord.com

Test Plan:
Tested clicking on an open thread. Then on closed threads to make sure the open thread loses focus. Also made sure that threads still hide behind the doc header and behind the ThreadedComments.

Before:

https://github.com/getcord/monorepo/assets/102963329/d1896aaa-9a69-4907-a692-71126910ad84


After:


https://github.com/getcord/monorepo/assets/102963329/ed1f87cd-766a-4ff6-b397-dadb30b9586a




Reviewers: km-nur

Reviewed By: km-nur

Pull Request: https://github.com/getcord/monorepo/pull/8086

monorepo-commit: 0e91c6aa73321750eaa09a92d95f2186809ee792